+++ examples/frontend-re-frame/src/clj/frontend_re_frame/core.clj
@@ -1,0 +1,1 @@
+(ns frontend-re-frame.core)