+++ examples/ring-swagger/src/example/server.clj
@@ -1,14 +1,17 @@
 (ns example.server
   (:require [reitit.ring :as ring]
+            [reitit.coercion.spec]
             [reitit.swagger :as swagger]
             [reitit.swagger-ui :as swagger-ui]
             [reitit.ring.coercion :as coercion]
-            [reitit.coercion.spec]
+            [reitit.dev.pretty :as pretty]
             [reitit.ring.middleware.muuntaja :as muuntaja]
             [reitit.ring.middleware.exception :as exception]
             [reitit.ring.middleware.multipart :as multipart]
             [reitit.ring.middleware.parameters :as parameters]
             [reitit.ring.middleware.dev :as dev]
+            [reitit.ring.spec :as spec]
+            [spec-tools.spell :as spell]
             [ring.adapter.jetty :as jetty]
             [muuntaja.core :as m]
             [clojure.java.io :as io]))
@@ -61,10 +64,15 @@
                             {:status 200
                              :body {:total (+ x y)}})}}]]]
 
-      {;;:reitit.middleware/transform dev/print-request-diffs
+      {;;:reitit.middleware/transform dev/print-request-diffs ;; pretty diffs
+       ;;:validate spec/validate ;; enable spec validation for route data
+       ;;:reitit.spec/wrap spell/closed ;; strict top-level validation
+       :exception pretty/exception
        :data {:coercion reitit.coercion.spec/coercion
               :muuntaja m/instance
-              :middleware [;; query-params & form-params
+              :middleware [;; swagger feature
+                           swagger/swagger-feature
+                           ;; query-params & form-params
                            parameters/parameters-middleware
                            ;; content-negotiation
                            muuntaja/format-negotiate-middleware